third_party/blink/web_tests/external/wpt/html/semantics/document-metadata/the-link-element/link-rel-attribute-tokenization-expected.txt

This is a testharness.js-based test.
[FAIL] The rel attribute needs to handle ASCII whitespace correctly
  assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
Harness: the test ran to completion.
